Restructure the application files.
