+++ README.md
@@ -26,6 +26,7 @@
 - setup a new postgresql DB, and/or get an existing live DB credentials
 - set a FALSK_ENV value: "development":  `export FLASK_ENV=development`
   (set "production" for the live server))
+- `export FLASK_APP=similarity`
 - `cp _example.config.py config.py`
 - set the relevant DB credentials in `config.py`
 - `flask db init` (initialize local db)
