+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='similarity',
+    packages=['similarity'],
+    include_package_data=True,
+    install_requires=[
+        'flask',
+    ],
+)
