+++ similarity/views.py
@@ -1,0 +1,6 @@
+from similarity import app
+
+
+@app.route('/')
+def hello_world():
+    return 'Hello, World!?'
